LSR-811 Merge and rename “ByArticleId” connector
